identity_profiling/requirements.txt
pandas
numpy
scikit-learn
tensorflow
keras
joblib
python-dotenv
matplotlib
seaborn
jupyter
